Completed part 2

+++ L1-spring-boot/src/main/java/com/example/demo/service/impl/BankManagementServiceImpl.java
@@ -49,10 +49,11 @@ public class BankManagementServiceImpl implements BankManagementService {
                 .toList();
     }
 
-    @Override
     @Transactional
     public void deleteUser(Long userId) {
-
+        if (!userRepository.existsById(userId)) { throw
+        new RuntimeException("User not found");
+        }
         userRepository.deleteById(userId);
     }
